Trait ndarray_npy::ViewMutElement

pub trait ViewMutElement: Sized {
    fn bytes_as_mut_slice<'a>(
        bytes: &'a mut [u8],
        type_desc: &PyValue,
        len: usize
    ) -> Result<&'a mut [Self], ViewDataError>;
}

An array element type that can be mutably viewed (without copying) in an .npy file.

Required Methods

fn bytes_as_mut_slice<'a>(
bytes: &'a mut [u8],
type_desc: &PyValue,
len: usize
) -> Result<&'a mut [Self], ViewDataError>

Casts bytes into a mutable slice of elements of length len.

Returns Err(_) in at least the following cases:

if the type_desc does not match Self with native endianness

if the bytes slice is misaligned for elements of type Self

if the bytes slice is too short for len elements

if the bytes slice has extra bytes after len elements

May panic if len * size_of::<Self>() overflows.